requirements.txt
requests
rich
gitpython